refactor: changes show more component
