+++ src/components/app/app.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {Router, Route, Switch} from 'react-router-dom';
 import history from '../../history';
 import {PathName} from '../../const';
-import Main from '../main/main.jsx';
+import MainPage from '../main-page/main-page.jsx';
 import MoviePage from '../movie-page/movie-page.jsx';
 
 const App = () => {
@@ -10,7 +10,7 @@ const App = () => {
     <Router history={history}>
       <Switch>
         <Route exact path={PathName.ROOT}>
-          <Main />
+          <MainPage />
         </Route>
         <Route path={`${PathName.MOVIE_PAGE}:id`}>
           {({match}) => (
