+++ src/const.js
@@ -2,7 +2,7 @@ const Config = {
   MOVIE_MAX_RATING: 10,
   MOVIE_ACTORS_TO_SHOW: 4,
   SIMILAR_MOVIES_TO_SHOW: 4,
-  MOVIES_NUMBER_TO_SHOW: 16,
+  MOVIES_NUMBER_TO_SHOW: 18,
   MOVIES_NUMBER_LIMIT: 8,
   MOVIES_NUMBER_OFFSET: 8,
   MAX_GENRES_TO_SHOW: 9,
@@ -14,7 +14,7 @@ const Config = {
     10: `Awesome`
   },
   ERRORS: {
-    FETCH_DATA: `Something went wrong, please try again later.`,
+    FETCH_DATA: `Something went wrong, please try again.`,
   },
   LOADER: {
     SMALL: `small`,
